midterm/requirements/dev.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai>=1.6.1
qdrant-client>=1.6.4
ragas==0.0.20
streamlit==1.29.0
python-dotenv==1.0.0
pypdf==3.17.1 